backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors==5.0.0
psycopg2-binary==2.9.9
stripe==10.5.0
python-dotenv==1.0.1
alembic==1.13.2
Werkzeug==3.0.4
gunicorn==22.0.0
openai>=1.0.0
reportlab==4.2.2
arabic-reshaper==3.0.0
python-bidi==0.4.2